Update global_functions_javascript.js
function showSummaryBox(divObject,parentLinkObj,showText,hideText) is called though OC’s code passing divObject as an actual element or the ID of an element.

The old function code had two different variable declarations to handle the two different calling scenarios:
     //var sumBox = $(divObject);
and
     var sumBox = document.getElementById(divObject);

Obviously this cannot work for both scenarios, without manually commenting and uncommenting the two lines (e.g. on page OpenClinica/SignStudySubject the link “Show events and discrepancy notes” is not functional).

The altered code is first checking to see if divObject is an element or its ID and then is declaring sumBox accordingly.