OC-6879 Pausing an extract job but it's still executed
Removing (pausing) an extract job changes the status of the job removed,
but the job still runs at the scheduled time.

Steps to Reproduce:

1. Create an Export Job & Schedule it to run at a specific time.
2. Navigate to Tasks > Administration > Jobs. Click "View all Export
Data Jobs"
3. Next to the job you created, click the "Remove" icon under the
"Actions" column.
4. Click "OK" when the confirmation message is displayed.
5. The status of the Job will now be removed.

Expected Results: The Job should not run at it's scheduled time until
the Job is restored.
Actual Results: The Job still runs at the expected time. This is
contrary to the confirmation message that is displayed to the user when
the job is removed.