OC-9767
Expose participant user's first name, email and phone number to runtime so that they could be used in the notification message .
This code is the call to UserService to return user list / study. For now , it will be on hold until we figure out how to create Roles to Participate users.